==> scripts/2_desc.Rmd <==
---
title: "Descriptive statistics"
author: "Alex Sheridan"
date: "2025-07-02"
output:
  html_document:
    css: styles.css
    toc: true
    toc_float: true
    toc_depth: 4
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE)

options(repos = c(CRAN = "https://cloud.r-project.org"))

for(
  pkg in c(
    "haven", # stata whatever
    "ggplot2", # graphs
    "viridis", # colors
    "dplyr", # basics
    "tidyr", # basics
    "survey", # survey weights
    "gtsummary", # summary tables
    "kableExtra", # styling of tables
    "kable",
    "beepr", # helps save time
    "stringr", # in pivoting: separate the suffix to get variables per year
    "data.table" # data.table 
  )
){
  if(!require(pkg, quietly = TRUE, character.only = TRUE)){
    install.packages(pkg)
  }
}

# Functions ####
source("scripts/0_functions.R")

# Data ####
load(file = "data/analysisdata/elfemini.Rdata")
load(file = "data/analysisdata/mcsmini.Rdata")

elfemini <- elfemini %>%
  mutate(across(-c(wgt1y, wgt2y, wgt3y, wgt5y), as.factor))

mcsmini <- mcsmini %>% 
  mutate(across(-c(wgt9m, wgt3y, wgt5y), as_factor))

for(i in c("sex")) {
  levels(elfemini[, i]) <- c("Boy", "Girl")
}

for(i in c("sex")) {
  levels(mcsmini[, i]) <- c("Boy", "Girl")
}

for(i in c("meduc3")) {
  levels(elfemini[, i]) <- c("Low", "Medium", "High")
}

for(i in c("meduc3")) {
  levels(mcsmini[, i]) <- c("Low", "Medium", "High")
}

for(i in c("meduc")) {
  levels(elfemini[, i]) <- c("<=bepc",  "cap-bep", "bac", "bac+2", "bac+3/4", ">bac+4")
}

for(i in c("meduc")) {
  levels(mcsmini[, i]) <- c("None", "Other", "GCSE D-G", "GCSE A-C", "Trade", "A-level", "HE below deg", "Bach", "Higher deg")
}

for(i in c("famstr2m", "famstr1y", "famstr2y", "famstr3y", "famstr5y")) {
  levels(elfemini[, i]) <- c("Two natural parents", "Mother only", "Father only", "Other")
}

for(i in c("famstr9m", "famstr3y", "famstr5y")) {
  levels(mcsmini[, i]) <- c("Two natural parents", "Mother only", "Father only", "Other")
}

for(i in c("twopar2m", "twopar1y", "twopar2y", "twopar3y", "twopar5y")) {
  levels(elfemini[, i]) <- c("No", "Yes")
}

for(i in c("twopar9m", "twopar3y", "twopar5y")) {
  levels(mcsmini[, i]) <- c("No", "Yes")
}

# Role modelling ####

for(i in c("nappies2m3", "tuckin2m3", "bath2m3", "walk2m3", "night2m3", "doctor2m3", "dishes2m3", "groceries2m3", "cook2m3", "laundry2m3", "clean2m3", "diy2m3")) {
  levels(elfemini[, i]) <- c("Mother mostly", "Balanced", "Father mostly")
}

for(i in c("nappies9m3", "night9m3", "cook9m3", "clean9m3", "laundry9m3", "diy9m3", "budgetting9m3", "doctor9m3", "lookafter9m3")) {
  levels(mcsmini[, i]) <- c("Mother mostly", "Balanced", "Father mostly")
}

for(i in c("emp2m", "emp1y", "emp2y", "emp3y", "emp5y")) {
  levels(elfemini[, i]) <- c("Both parents", "Father only", "Mother only", "Neither")
}

for(i in c("emp9m", "emp3y")) {
  levels(mcsmini[, i]) <- c("Both parents", "Father only", "Mother only", "Neither")
}

for(i in c("mpartemp2m", "fpartemp2m", "mpartemp1y", "fpartemp1y", "mpartemp2y", "fpartemp2y", "mpartemp3y", "fpartemp3y")) {
  levels(elfemini[, i]) <- c("No", "Yes")
}

for(i in c("mpartemp9m", "fpartemp9m", "mpartemp3y")) {
  levels(mcsmini[, i]) <- c("No", "Yes")
}

# Attitudes expectations ####

for(i in c("socialsuccess2m", "lovelife2m", "interestingjob2m", "passion2m", "calmlife2m", "bigfamily2m", "lotsoffriends2m", "fairerworld2m", "goodhealth2m", "otherwish2m")) {
  levels(elfemini[, i]) <- c("No", "Yes")
}

for(i in c("independence3y", "obedience3y", "negotiation3y", "respectelders3y", "dowellatschool3y", "instillreligiousvalues3y",
           "bewellliked3y", "thinkforself3y", "workhard3y", "helpothers3y", "obeyparents3y", "qualityreligiousvalues3y")) {
  levels(mcsmini[, i]) <- c("No", "Yes")
}

# Children's access to resources ####

for(i in c("frpaint3y", "frread3y", "music3y", "readplus3y", "frcounting3y", "writing3y", "puzzle3y",
           "anyactivity3y",
           "swimming3y", "gymnastics3y", "circus3y", "sportsinit3y", "musicclass3y", "danceclass3y", "visualarts3y", "horseriding3y")) {
  levels(elfe5_fact[, i]) <- c("No", "Yes")
}

for(i in c("read3y", "library3y", "counting3y", "songs3y", "alphabet3y", "paint3y", "physical3y",
           "read5y", "stories5y", "songs5y", "paint5y", "physical5y", "indoor5y", "park5y")) {
  levels(mcs5_fact[, i]) <- c("No", "Yes")
}

for(i in c("fqcounting3y", "fqsongs3y", "fqalphabet3y", "fqpaint3y")) {
  levels(mcs5_fact[, i]) <- c("Occasionally or less than once a week", "1 - 2 days per week", "3 times a week", "4 times a week", "5 times a week", "6 times a week", "7 times a week constantly")
}

for(i in c("fqlibrary3y")) {
  levels(mcs5_fact[, i]) <- c("On special occasions", "Once a month", "Once a fortnight", "Once a week")
}

for(i in c("fqread3y", "fqread5y", "fqstories5y", "fqsongs5y", "fqpaint5y", "fqphysical5y", "fqindoor5y", "fqpark5y")) {
  levels(mcs5_fact[, i]) <- c("Not at all", "Less often", "Once or twice a month", "Once or twice a week", "Several times a week", "Every day")
}

elfe5 <- elfemini %>%
  drop_na(wgt5y)
welfe5 <- svydesign(ids = ~1, data = elfe5, weights = ~ elfe5$wgt5y)

mcs5 <- mcsmini %>%
  filter(inwave5y == 1)
wmcs5 <- svydesign(ids = ~1, data = mcs5, weights = ~ mcs5$wgt5y)

# Role modelling sample 

elfewhodoes <- elfe5 %>% 
  drop_na(doctor2m3, nappies2m3, laundry2m3, night2m3, clean2m3, cook2m3, diy2m3)
welfewhodoes <- svydesign(ids = ~1, data = elfewhodoes, weights = ~ elfewhodoes$wgt5y)

mcswhodoes <- mcs5 %>%
  drop_na(doctor9m3, nappies9m3, laundry9m3, night9m3, clean9m3, cook9m3, diy9m3)
wmcswhodoes <- svydesign(ids = ~1, data = mcswhodoes, weights = ~ mcswhodoes$wgt5y)

# Attitudes expectations sample 

elfeexpect <- elfe5 %>% 
  drop_na(socialsuccess2m, lovelife2m, interestingjob2m, passion2m, calmlife2m, bigfamily2m, lotsoffriends2m, fairerworld2m, goodhealth2m, otherwish2m)
welfeexpect <- svydesign(ids = ~1, data = elfeexpect, weights = ~ elfeexpect$wgt5y)

mcsexpect <- mcs5 %>%
  drop_na(independence3y, obedience3y, negotiation3y, respectelders3y, dowellatschool3y, instillreligiousvalues3y,
           bewellliked3y, thinkforself3y, workhard3y, helpothers3y, obeyparents3y, qualityreligiousvalues3y)
wmcsexpect <- svydesign(ids = ~1, data = mcsexpect, weights = ~ mcsexpect$wgt5y)

```

## Gender, education, family structure {.tabset}

### France

```{r}
welfe5 %>% tbl_svysummary(include = c("sex",
                                      "meduc",
                                      "meduc3"),
                              label = list(sex ~ "Child's sex",
                                           meduc ~ "Mother's education",
                                           meduc3 ~ "Mother's education (3 cat)"),
                              statistic = all_categorical() ~ c("{p}"),
                              digits = list( ~ c(1, 1)),
                              missing = "no") %>%
  modify_header(label = "**Variable**", all_stat_cols() ~ "**All (%)**, N = {n_unweighted}") %>%
  add_n("{N_miss_unweighted}", col_label = "**Missings**") %>%
  modify_caption("Child's sex and mother's education in France") %>%
  as_kable_extra(booktabs = TRUE, linesep = "", longtable = TRUE) %>%
  kable_styling(latex_options = c("HOLD_position", "repeat_header"))


# Reshape and prepare data
long_data <- elfe5 %>%
  select(famstr2m, famstr1y, famstr2y, famstr3y, famstr5y) %>%
  pivot_longer(cols = everything(), 
               names_to = "time_point", 
               values_to = "family_structure") %>%
  mutate(time = factor(str_extract(time_point, "\\d+[my]"), levels = c("5y", "3y", "2y", "1y", "2m")))

# Calculate percentages
percent_data <- long_data %>%
  group_by(time, family_structure) %>%
  summarise(n = n(), .groups = "drop") %>%
  group_by(time) %>%
  mutate(percent = n / sum(n))

# Plot
ggplot(percent_data, aes(x = time, y = percent, fill = family_structure)) +
  geom_bar(stat = "identity", position = "stack") +
  coord_flip() +
  scale_fill_manual(values = rev(viridis(4, option = "D"))) +
  scale_y_continuous(labels = scales::percent_format()) +
  labs(title = "Family structure in France, from 2 months to age 5",
       subtitle = "By age 5, around 87% of children still live with both natural parents",
       x = "Wave",
       y = "Percentage",
       fill = "Family Structure") +
  theme_minimal()

```


### UK

```{r}
wmcs5 %>% tbl_svysummary(include = c("sex",
                                      "meduc",
                                      "meduc3"),
                              label = list(sex ~ "Child's sex",
                                           meduc ~ "Mother's education",
                                           meduc3 ~ "Mother's education (3 cat)"),
                              statistic = all_categorical() ~ c("{p}"),
                              digits = list( ~ c(1, 1)),
                              missing = "no") %>%
  modify_header(label = "**Variable**", all_stat_cols() ~ "**All (%)**, N = {n_unweighted}") %>%
  add_n("{N_miss_unweighted}", col_label = "**Missings**") %>%
  modify_caption("Child's sex and mother's education in the UK") %>%
  as_kable_extra(booktabs = TRUE, linesep = "", longtable = TRUE) %>%
  kable_styling(latex_options = c("HOLD_position", "repeat_header"))



# Reshape and prepare data
long_data <- mcs5 %>%
  select(famstr9m, famstr3y, famstr5y) %>%
  pivot_longer(cols = everything(), 
               names_to = "time_point", 
               values_to = "family_structure") %>%
  mutate(time = factor(str_extract(time_point, "\\d+[my]"), levels = c("5y", "3y", "9m")))

# Calculate percentages
percent_data <- long_data %>%
  group_by(time, family_structure) %>%
  summarise(n = n(), .groups = "drop") %>%
  group_by(time) %>%
  mutate(percent = n / sum(n))

# Plot
ggplot(percent_data, aes(x = time, y = percent, fill = family_structure)) +
  geom_bar(stat = "identity", position = "stack") +
  coord_flip() +
  scale_fill_manual(values = rev(viridis(4, option = "D"))) +
  scale_y_continuous(labels = scales::percent_format()) +
  labs(title = "Family structure in the UK, from 9 months to age 5",
       subtitle = "By age 5, around 75% of children still live with both natural parents",
       x = "Wave",
       y = "Percentage",
       fill = "Family Structure") +
  theme_minimal()


```

## Role modelling {.tabset}

### France

```{r}

welfe5 %>% tbl_svysummary(include = c("doctor2m3", "nappies2m3", "laundry2m3", "night2m3", "clean2m3", "cook2m3", "diy2m3"),
                          label = list(),
                          statistic = all_categorical() ~ c("{p}"),
                          digits = list( ~ c(1, 1)),
                          by = "sex",
                          missing = "no") %>%
  modify_header(label = "**Variable**", all_stat_cols() ~ "**{level}**, N = {n_unweighted}") %>%
  add_p() %>% 
  add_n("{N_miss_unweighted}", col_label = "**Missings**") %>%
  modify_caption("Who does what at home? France") %>%
  as_kable_extra(booktabs = TRUE, linesep = "", longtable = TRUE) %>%
  kable_styling(latex_options = c("HOLD_position", "repeat_header"))


welfe5 %>% tbl_svysummary(include = c("emp2m", "emp1y", "emp2y", "emp3y", "emp5y"),
                          label = list(),
                          statistic = all_categorical() ~ c("{p}"),
                          digits = list( ~ c(1, 1)),
                          by = "sex",
                          missing = "no") %>%
  modify_header(label = "**Variable**", all_stat_cols() ~ "**{level}**, N = {n_unweighted}") %>%
  add_p() %>% 
  add_n("{N_miss_unweighted}", col_label = "**Missings**") %>%
  modify_caption("Who works outside the home? France") %>%
  as_kable_extra(booktabs = TRUE, linesep = "", longtable = TRUE) %>%
  kable_styling(latex_options = c("HOLD_position", "repeat_header"))

welfe5 %>% tbl_svysummary(include = c("mpartemp2m", "fpartemp2m", "mpartemp1y", "fpartemp1y", "mpartemp2y", "fpartemp2y", "mpartemp3y", "fpartemp3y"),
                          label = list(),
                          statistic = all_categorical() ~ c("{p}"),
                          digits = list( ~ c(1, 1)),
                          by = "sex",
                          missing = "no") %>%
  modify_header(label = "**Variable**", all_stat_cols() ~ "**{level}**, N = {n_unweighted}") %>%
  add_p() %>% 
  add_n("{N_miss_unweighted}", col_label = "**Missings**") %>%
  modify_caption("Who works part-time? France") %>%
  as_kable_extra(booktabs = TRUE, linesep = "", longtable = TRUE) %>%
  kable_styling(latex_options = c("HOLD_position", "repeat_header"))

```


### UK

```{r}

wmcs5 %>% tbl_svysummary(include = c("doctor9m3", "nappies9m3", "laundry9m3", "night9m3", "clean9m3", "cook9m3", "diy9m3"),
                        label = list(),
                        statistic = all_categorical() ~ c("{p}"),
                        digits = list( ~ c(1, 1)),
                        by = "sex",
                        missing = "no") %>%
  modify_header(label = "**Variable**", all_stat_cols() ~ "**{level}**, N = {n_unweighted}") %>%
  add_p() %>% 
  add_n("{N_miss_unweighted}", col_label = "**Missings**") %>%
  modify_caption("Who does what at home? UK") %>%
  as_kable_extra(booktabs = TRUE, linesep = "", longtable = TRUE) %>%
  kable_styling(latex_options = c("HOLD_position", "repeat_header"))


wmcs5 %>% tbl_svysummary(include = c("emp9m", "emp3y"),
                        label = list(),
                        statistic = all_categorical() ~ c("{p}"),
                        digits = list( ~ c(1, 1)),
                        by = "sex",
                        missing = "no") %>%
  modify_header(label = "**Variable**", all_stat_cols() ~ "**{level}**, N = {n_unweighted}") %>%
  add_p() %>% 
  add_n("{N_miss_unweighted}", col_label = "**Missings**") %>%
  modify_caption("Who works outside the home? UK") %>%
  as_kable_extra(booktabs = TRUE, linesep = "", longtable = TRUE) %>%
  kable_styling(latex_options = c("HOLD_position", "repeat_header"))

wmcs5 %>% tbl_svysummary(include = c("mpartemp9m", "fpartemp9m", "mpartemp3y"),
                        label = list(),
                        statistic = all_categorical() ~ c("{p}"),
                        digits = list( ~ c(1, 1)),
                        by = "sex",
                        missing = "no") %>%
  modify_header(label = "**Variable**", all_stat_cols() ~ "**{level}**, N = {n_unweighted}") %>%
  add_p() %>% 
  add_n("{N_miss_unweighted}", col_label = "**Missings**") %>%
  modify_caption("Who works part-time? UK") %>%
  as_kable_extra(booktabs = TRUE, linesep = "", longtable = TRUE) %>%
  kable_styling(latex_options = c("HOLD_position", "repeat_header"))

```

## Attitudes, expectations {.tabset}

### France

```{r}
welfe5 %>% tbl_svysummary(include = c("socialsuccess2m", "lovelife2m", "interestingjob2m", "passion2m", "calmlife2m", "bigfamily2m", "lotsoffriends2m", "fairerworld2m", "goodhealth2m", "otherwish2m"),
                          label = list(),
                          statistic = all_categorical() ~ c("{p}"),
                          digits = list( ~ c(1, 1)),
                          by = "sex",
                          missing = "no") %>%
  modify_header(label = "**Variable**", all_stat_cols() ~ "**{level}**, N = {n_unweighted}") %>%
  add_p() %>% 
  add_n("{N_miss_unweighted}", col_label = "**Missings**") %>%
  modify_caption("What do parents wish for their child? France") %>%
  as_kable_extra(booktabs = TRUE, linesep = "", longtable = TRUE) %>%
  kable_styling(latex_options = c("HOLD_position", "repeat_header"))
```

### UK

```{r}
wmcs5 %>% tbl_svysummary(include = c("independence3y", "obedience3y", "negotiation3y", "respectelders3y", "dowellatschool3y", "instillreligiousvalues3y", "bewellliked3y", "thinkforself3y", "workhard3y", "helpothers3y", "obeyparents3y", "qualityreligiousvalues3y"),
                        label = list(),
                        statistic = all_categorical() ~ c("{p}"),
                        digits = list( ~ c(1, 1)),
                        by = "sex",
                        missing = "no") %>%
  modify_header(label = "**Variable**", all_stat_cols() ~ "**{level}**, N = {n_unweighted}") %>%
  add_p() %>% 
  add_n("{N_miss_unweighted}", col_label = "**Missings**") %>%
  modify_caption("What do parents wish for their child? UK") %>%
  as_kable_extra(booktabs = TRUE, linesep = "", longtable = TRUE) %>%
  kable_styling(latex_options = c("HOLD_position", "repeat_header"))


```

## Children's access to resources {.tabset}

### France

```{r}
welfe5 %>% tbl_svysummary(include = c("frpaint3y", "frread3y", "music3y", "readplus3y", "frcounting3y", "writing3y", "puzzle3y", "anyactivity3y", "swimming3y", "gymnastics3y", "circus3y", "sportsinit3y", "musicclass3y", "danceclass3y", "visualarts3y", "horseriding3y"),
                          label = list(),
                          statistic = all_categorical() ~ c("{p}"),
                          digits = list( ~ c(1, 1)),
                          by = "sex",
                          missing = "no") %>%
  modify_header(label = "**Variable**", all_stat_cols() ~ "**{level}**, N = {n_unweighted}") %>%
  add_p() %>% 
  add_n("{N_miss_unweighted}", col_label = "**Missings**") %>%
  modify_caption("What do parents do with their children? What activities do they choose? France") %>%
  as_kable_extra(booktabs = TRUE, linesep = "", longtable = TRUE) %>%
  kable_styling(latex_options = c("HOLD_position", "repeat_header"))
```

### UK

```{r}

wmcs5 %>% tbl_svysummary(include = c("read3y", "library3y", "counting3y", "songs3y", "alphabet3y", "paint3y", "physical3y",
           "read5y", "stories5y", "songs5y", "paint5y", "physical5y", "indoor5y", "park5y",
           "fqread3y", "fqcounting3y", "fqsongs3y", "fqalphabet3y", "fqpaint3y",
           "fqlibrary3y", 
           "fqread5y", "fqstories5y", "fqsongs5y", "fqpaint5y", "fqphysical5y", "fqindoor5y", "fqpark5y"),
                        label = list(),
                        statistic = all_categorical() ~ c("{p}"),
                        digits = list( ~ c(1, 1)),
                        by = "sex",
                        missing = "no") %>%
  modify_header(label = "**Variable**", all_stat_cols() ~ "**{level}**, N = {n_unweighted}") %>%
  add_p() %>% 
  add_n("{N_miss_unweighted}", col_label = "**Missings**") %>%
  modify_caption("What do parents do with their children? What activities do they choose? UK") %>%
  as_kable_extra(booktabs = TRUE, linesep = "", longtable = TRUE) %>%
  kable_styling(latex_options = c("HOLD_position", "repeat_header"))

```
